RC_Channel: split RC_Channels.cpp out to separate cpp file
this is needed for automatic parameter documentation generation
